BugSuggestion components move to components/feedback directory
