+++ src/components/Ticketbox/components/Feedback/BugSuggestion.jsx
@@ -1,11 +1,12 @@
 import * as React from "react";
-import { DynamicFormContainer } from "../../DynamicForm";
+import { DynamicFormContainer } from "../../../DynamicForm";
 import { gql } from "apollo-boost";
-import { client } from "../../../";
-import { QA } from './SuggestionBugQA.js';
-import Success from './Success';
-import Error from './Error';
-import BackBtn from './BackBtn';
+import { client } from "../../../../";
+import { QA } from './BugSuggestionQA';
+import Success from '../Success';
+import Error from '../Error';
+import BackBtn from '../BackBtn';
+
 class BugSuggestion extends React.Component {
   state = { error: null, response: null };
 
@@ -38,7 +39,7 @@ class BugSuggestion extends React.Component {
     const { category } = this.props;
     const { error, response } = this.state;
     const imgFile = category === 'bug' ? 'Artboard 3-small.png' : 'Artboard 2-small.png';
-    const imgSrc = require(`../../../assets/${imgFile}`);
+    const imgSrc = require(`../../../../assets/${imgFile}`);
 
     if (error) {
       return (
