+++ src/components/Ticketbox/components/Feedback/index.js
@@ -1,0 +1,5 @@
+import BugSuggestion from "./BugSuggestion.jsx";
+
+export {
+  BugSuggestion,
+};
